+++ SCF/Assignment2/FCFS/src/FCFS.java
@@ -1,8 +1,6 @@
-import java.util.Scanner;
-
-/*
+/**
  * @author ${kunal kaushik}
- * desc : This assignment is of First Come First Serve (FCFS)
+ * @desc : This assignment is of First Come First Serve (FCFS)
  *        array size is Dynamically taken at run time
  *        Here all the process say n are stored in 2-D array (process[n][5])
  *        where for any process sai 'i',
@@ -14,40 +12,12 @@ import java.util.Scanner;
  *  
  */
 public class FCFS {
-	private static Scanner sc = new Scanner(System.in);
-
-	public static void main(String argv[]) {
-		int process[][]; // this array stores the info. of all the processes
-		try {
-			System.out.println("Enter the number of process : ");
-			int n = sc.nextInt();// no. of process
-			process = new int[n][5]; // dynamic input array
-			for (int index = 0; index < n; index++) {
-				System.out.println("Enter the Arrival time for process["
-						+ index + "] : ");
-				process[index][0] = sc.nextInt(); // ArrivalTime
-				System.out.println("Enter the Burst time for process[" + index
-						+ "] : ");
-				process[index][1] = sc.nextInt(); // BurstTime
-				System.out
-						.println("..............................................\n");
-			}
-
-			FCFS.completionTime(process);
-			FCFS.turnAroundTime(process);
-			FCFS.waitingTime(process);
-			FCFS.display(process);
-		} catch (Exception ex) // If invalid input is entered
-		{
-			System.out.println("Wrong input...\nTry Again later..");
-		}
-	}
-
-	/*
-	 * auther : Kunal Kaushik
-	 * desc : this method calculates waitingTime for each process
-	 * params : 2-D array of int type which stores info. of each process
-	 * returnType : void
+	
+	/**
+	 * @author : Kunal Kaushik
+	 * @desc : this method calculates waitingTime for each process
+	 * @params : 2-D array of int type which stores info. of each process
+	 * @return : void
 	 */
 	public static void waitingTime(int process[][]) {
 		for (int index = 0; index < process.length; index++) {
@@ -57,11 +27,11 @@ public class FCFS {
 
 	}
 
-	/*
-	 * auther : Kunal Kaushik 
-	 * desc : this method calculates turnAroundTime for each process
-	 * params : 2-D array of int type which stores info. of each process
-	 * returnType : void
+	/**
+	 * @author : Kunal Kaushik 
+	 * @desc : this method calculates turnAroundTime for each process
+	 * @params : 2-D array of int type which stores info. of each process
+	 * @return : void
 	 */
 	public static void turnAroundTime(int process[][]) {
 
@@ -70,11 +40,11 @@ public class FCFS {
 			process[index][3] = process[index][4] - process[index][0]; // turnAroundTime
 		}
 	}
-	/*
-	 * auther : Kunal Kaushik 
-	 * desc : this method calculates completionTime for each process
-	 * params : 2-D array of int type which stores info. of each process
-	 * returnType : void
+	/**
+	 * @author : Kunal Kaushik 
+	 * @desc : this method calculates completionTime for each process
+	 * @params : 2-D array of int type which stores info. of each process
+	 * @return : void
 	 */
 	public static void completionTime(int process[][]) {
 
@@ -102,11 +72,11 @@ public class FCFS {
 			}
 		}
 	}
-	/*
-	 * auther : Kunal Kaushik 
-	 * desc : this method calculates averageWaitingTime of processes
-	 * params : 2-D array of int type which stores info. of each process
-	 * returnType : double i.e, returns the averageWaitingTime
+	/**
+	 * @author : Kunal Kaushik 
+	 * @desc : this method calculates averageWaitingTime of processes
+	 * @params : 2-D array of int type which stores info. of each process
+	 * @return : double i.e, returns the averageWaitingTime
 	 */
 	public static double avgWaitingTime(int process[][]) {
 		double avgWaitTime = 0;
@@ -114,11 +84,11 @@ public class FCFS {
 			avgWaitTime += process[index][2];
 		return (avgWaitTime / process.length); // average Waiting Time
 	}
-	/*
-	 * auther : Kunal Kaushik 
-	 * desc : this method calculates averageTurnAroundTime of processes
-	 * params : 2-D array of int type which stores info. of each process
-	 * returnType : double i.e, returns averageTurnAroundTime
+	/**
+	 * @author : Kunal Kaushik 
+	 * @desc : this method calculates averageTurnAroundTime of processes
+	 * @params : 2-D array of int type which stores info. of each process
+	 * @return : double i.e, returns averageTurnAroundTime
 	 */
 	public static double avgTurnAroundTime(int process[][]) {
 		double avgTurnAroundTime = 0;
@@ -126,11 +96,11 @@ public class FCFS {
 			avgTurnAroundTime += process[index][3];
 		return (avgTurnAroundTime / process.length); // average TurnAroundTime
 	}
-	/*
-	 * auther : Kunal Kaushik 
-	 * desc : this method displays all the info. for each process
-	 * params : 2-D array of int type which stores info. of each process
-	 * returnType : void
+	/**
+	 * @author : Kunal Kaushik 
+	 * @desc : this method displays all the info. for each process
+	 * @params : 2-D array of int type which stores info. of each process
+	 * @return : void
 	 */
 	public static void display(int process[][]) {
 		System.out
