+++ SCF/Assignment2/FCFS/src/FcfsExecute.java
@@ -1,0 +1,38 @@
+import java.util.Scanner;
+/**
+ * 
+ * @author : Kunal
+ *@desc    : this class uses predefined functions of class FCFS 
+ */
+
+public class FcfsExecute {
+	private static Scanner sc = new Scanner(System.in);
+
+	public static void main(String argv[]) {
+		int process[][]; // this array stores the info. of all the processes
+		try {
+			System.out.println("Enter the number of process : ");
+			int n = sc.nextInt();// no. of process
+			process = new int[n][5]; // dynamic input array
+			for (int index = 0; index < n; index++) {
+				System.out.println("Enter the Arrival time for process["
+						+ index + "] : ");
+				process[index][0] = sc.nextInt(); // ArrivalTime
+				System.out.println("Enter the Burst time for process[" + index
+						+ "] : ");
+				process[index][1] = sc.nextInt(); // BurstTime
+				System.out
+						.println("..............................................\n");
+			}
+
+			FCFS.completionTime(process);
+			FCFS.turnAroundTime(process);
+			FCFS.waitingTime(process);
+			FCFS.display(process);
+		} catch (Exception ex) // If invalid input is entered
+		{
+			System.out.println("Wrong input...\nTry Again later..");
+		}
+	}
+
+}
